requirements.txt
pandas>=1.3.5
numpy>=1.21.6
boto3>=1.26.29
fastparquet>=0.8.1
pyarrow>=11.0.0
python-dotenv